+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.6.0 (26.10.2022)
+
+* Added [Kadena Chainweb Client](https://github.com/kommitters/kadena.ex/issues/59) functions
+* Added [Chainweb Pact Behaviour](https://github.com/kommitters/kadena.ex/issues/96)
+* Added [Error struct](https://github.com/kommitters/kadena.ex/pull/111) for the `Chainweb.Client` boundary
+
 ## 0.5.0 (21.10.2022)
 
 * Added [Kadena Pact](https://github.com/kommitters/kadena.ex/issues/55) functions